refactor: split tools.jsx into themed tool modules
Break the 1016-line tools.jsx (over the 800-line project max) into themed
component files plus non-component helpers/constants/registry modules. Each
file now exports either only React components or only non-components, which
fixes all 33 react-refresh/only-export-components errors and restores clean
Vite fast-refresh for toolbar edits.

No component logic changes — pure cut/paste + import wiring. The three
importers (Toolbar.jsx, ToolbarGroup.jsx, tools.test.js) repoint from
./tools to ./toolRegistry; registry exports (TOOL_DEFINITIONS, CORE_GROUPS,
EXTRA_GROUPS) stay byte-identical in shape.
